Weekly Update
- weapon only one now
+ weapon data: attack speed, cooldown
+ upgrade system: main logic is whenever kill a enemy, it will upgrade once, upgrade condition is just for now.
+ fixed attack: the attack logic now is calculate the distance of the mouse and the distance of the enemy, using attack range to to see if attack in range or not, then kill.

+++ Chef Strikes Back/Assets/Scripts/Enemy/Enemy.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 
 public class Enemy : MonoBehaviour
 {
-    public int health = 10;
+    public float health = 10.0f;
 
-    public void TakeDamage(int damage)
+    public void TakeDamage(float damage)
     {
         health -= damage;
 
-        if (health <= 0)
+        if (health <= 0.0f)
         {
             Destroy(gameObject);
             Debug.Log("Enemy destroyed.");
